refactor: separate profile research runtime responsibilities
Changed:
Separated Redis sessions, MarkdownFlow validation, replay events, and the shared LLM provider from the profile research orchestrator.

Benefit:
Profile collection keeps the same wire and persistence behavior while each runtime responsibility can be understood and tested independently.
